ai-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
numpy>=1.26.4
opencv-python-headless>=4.9.0.80
torch>=2.2.2
torchvision>=0.17.2
Pillow>=10.3.0
timm>=1.0.3
mediapipe>=0.10.14
aiofiles>=23.2.1
python-dotenv>=1.0.1
scipy>=1.13.0
albumentations==1.3.1
scikit-learn>=1.4.0
open_clip_torch>=2.24.0
tensorboard>=2.16.0
datasets>=2.14.0
